// src/app/core/models/fixture.type.ts
export interface FixtureType {
    ending_at: string;
    finished: boolean;
    fixtures: any;
    games_in_current_week: boolean;
    id: number;
    is_current: boolean;
    league_id: number;
    name: string;
    season_id: number;
    sport_id: number;
    stage_id: number;
    starting_at: string;
  }